extend json, setup loader
